+++ packages/lumx-react/src/components/slideshow/Slideshow.stories.tsx
@@ -1,6 +1,7 @@
-import { AspectRatio, ImageBlock, Slideshow, SlideshowItem, SlideshowProps } from '@lumx/react';
-import { boolean, number } from '@storybook/addon-knobs';
 import React from 'react';
+import range from 'lodash/range';
+import { AspectRatio, Button, FlexBox, ImageBlock, Slideshow, SlideshowItem } from '@lumx/react';
+import { boolean, number } from '@storybook/addon-knobs';
 import { thumbnailsKnob } from '@lumx/react/stories/knobs/thumbnailsKnob';
 
 export default { title: 'LumX components/slideshow/Slideshow' };
@@ -12,16 +13,15 @@ export const Simple = ({ theme }: any) => {
     const autoPlay = boolean('Autoplay', false);
     const interval = number('Autoplay interval (in milliseconds)', 1000);
 
-    const slideshowControlsProps: SlideshowProps['slideshowControlsProps'] = {
-        nextButtonProps: { label: 'Next' },
-        previousButtonProps: { label: 'Previous' },
-    };
     return (
         <Slideshow
             activeIndex={activeIndex}
             autoPlay={autoPlay}
             interval={interval}
-            slideshowControlsProps={slideshowControlsProps}
+            slideshowControlsProps={{
+                nextButtonProps: { label: 'Next' },
+                previousButtonProps: { label: 'Previous' },
+            }}
             theme={theme}
             groupBy={groupBy}
             style={{ width: '50%' }}
@@ -37,5 +37,45 @@ export const Simple = ({ theme }: any) => {
                 </SlideshowItem>
             ))}
         </Slideshow>
+    );
+};
+
+export const ResponsiveSlideShowSwipe = () => {
+    const slides = range(3);
+    return (
+        <>
+            In responsive mode
+            <ul>
+                <li>The slideshow is swipe-able (horizontal left/right)</li>
+                <li>The vertical scroll should still be possible</li>
+                <li>Clicking on elements inside a slide should work correctly</li>
+            </ul>
+            <FlexBox vAlign="center">
+                <Slideshow
+                    slideshowControlsProps={{
+                        nextButtonProps: { label: 'Next' },
+                        previousButtonProps: { label: 'Previous' },
+                    }}
+                >
+                    {slides.map((slide) => (
+                        <SlideshowItem key={`${slide}`}>
+                            <FlexBox
+                                style={{ border: '1px solid grey', maxWidth: 300, height: 300 }}
+                                hAlign="center"
+                                vAlign="center"
+                            >
+                                <Button onClick={() => alert(`Clicked button ${slide}`)}>Button {slide}</Button>
+                            </FlexBox>
+                        </SlideshowItem>
+                    ))}
+                </Slideshow>
+            </FlexBox>
+            {
+                /* Line padding to test the vertical scroll.*/
+                range(100).map((i) => (
+                    <br key={i} />
+                ))
+            }
+        </>
     );
 };